[FIX/#99] 공고 네비게이션 수정

+++ data/src/main/java/com/terning/data/dto/request/ScrapRequestDto.kt
@@ -7,7 +7,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class ScrapRequestDto(
     @SerialName("internshipAnnouncementId")
-    val internshipAnnouncementId: Int,
+    val internshipAnnouncementId: Long,
     @SerialName("color")
     val color: Int,
 )
