+++ domain/src/main/java/com/terning/domain/entity/request/ScrapRequestModel.kt
@@ -1,6 +1,6 @@
 package com.terning.domain.entity.request
 
 data class ScrapRequestModel(
-    val internshipAnnouncementId: Int,
+    val internshipAnnouncementId: Long,
     val color: Int,
 )
